add description quality to app

+++ app/api/infrastructure/graph/model/models_gen.go
@@ -40,7 +40,7 @@ type InternalRepository struct {
 }
 
 type QualityProjectDesc struct {
-	Quality int `json:"quality"`
+	Quality float64 `json:"quality"`
 }
 
 type RemoveAPITokenInput struct {
